Cleanup proto helpers and change ReadContract result to base64 string
